requirements.txt
colorama
dnspython
requests
tqdm